+++ .github/workflows/test.yml
@@ -10,4 +10,5 @@ jobs:
           node-version: 16
           registry-url: 'https://registry.npmjs.org'
       - run: npm ci
+      - run: npm run build
       - run: npm test
