+++ .gitignore
@@ -2,3 +2,5 @@
 
 node_modules
 yarn-error.log
+
+dist
