+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ecoindex",
-  "version": "1.0.3",
+  "version": "1.0.4@beta",
   "main": "index.js",
   "repository": "https://github.com/tsecher/ecoindex_js.git",
   "author": "tsecher <me@thomas-secher.fr>",
@@ -10,9 +10,13 @@
     "node": ">=16.0.0"
   },
   "scripts": {
-    "test": "mocha"
+    "test": "mocha",
+    "build": "node scripts/build.js"
   },
   "devDependencies": {
+    "@babel/core": "^7.22.1",
+    "@babel/plugin-syntax-import-attributes": "^7.22.3",
+    "@babel/preset-env": "^7.22.4",
     "chai": "^4.3.7",
     "mocha": "^10.2.0"
   },
@@ -21,8 +25,8 @@
   },
   "exports": {
     ".": {
-      "import": "./index.js",
-      "require": "./index.cjs"
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs"
     }
   }
 }
